+++ src/App.tsx
@@ -25,72 +25,12 @@ import {
   isUnitHero,
 } from "./utils";
 
-/**
- * https://army-forge.onepagerules.com/api/rules/common/1 is ???
- * https://army-forge.onepagerules.com/api/rules/common/2 is grimdark future
- * https://army-forge.onepagerules.com/api/rules/common/3 is grimdark future firefight
- * https://army-forge.onepagerules.com/api/rules/common/4 is age of fantasy
- * https://army-forge.onepagerules.com/api/rules/common/5 is age of fantasy skirmish
- * https://army-forge.onepagerules.com/api/rules/common/6 is age of fantasy: regiments
- * https://army-forge.onepagerules.com/api/rules/common/7 is age of fantasy: quest
- * https://army-forge.onepagerules.com/api/rules/common/9 is grimdark future: star quest
- * warfleets doesnt seem to have one?
- */
-
 import { ArrowPath, Cross, Duplicate, Cog } from "./components/icons";
 import { DarkModeSwitch } from "./components/DarkModeSwitch";
-import { useEffect, useState } from "react";
 import { LanguagePicker } from "./components/LanguagePicker";
-
-const gameSystemMappingCommonRules = {
-  "grimdark-future": 2,
-  "grimdark-future-firefight": 3,
-  "age-of-fantasy": 4,
-  "age-of-fantasy-skirmish": 5,
-  "age-of-fantasy-regiments": 6,
-} as const;
 
 function App() {
   const stateView = useSnapshot(state, { sync: true });
-
-  const [commonRules, setCommonRules] = useState({
-    [gameSystemMappingCommonRules["grimdark-future"]]: [],
-    [gameSystemMappingCommonRules["grimdark-future-firefight"]]: [],
-    [gameSystemMappingCommonRules["age-of-fantasy"]]: [],
-    [gameSystemMappingCommonRules["age-of-fantasy-skirmish"]]: [],
-    [gameSystemMappingCommonRules["age-of-fantasy-regiments"]]: [],
-  });
-
-  useEffect(() => {
-    const x = async () => {
-      (
-        Object.keys(
-          gameSystemMappingCommonRules
-        ) as (keyof typeof gameSystemMappingCommonRules)[]
-      ).forEach(async (gameSystem) => {
-        const id = gameSystemMappingCommonRules[gameSystem];
-        const response = await fetch(
-          `https://army-forge.onepagerules.com/api/rules/common/${id}`
-        );
-        const data = await response.json();
-
-        setCommonRules((prev) => ({
-          ...prev,
-          [gameSystemMappingCommonRules[gameSystem]]: data,
-        }));
-      });
-
-      const response = await fetch(
-        `https://army-forge.onepagerules.com/api/rules/common/${gameSystemMappingCommonRules["grimdark-future"]}`
-      );
-      const data = await response.json();
-      if (response.status !== 200) {
-        alert("Army Forge failed to export list. Sorry!");
-      }
-    };
-
-    // x();
-  }, []);
 
   const areAllLoadoutsChecked = stateView.unitProfiles.every((unit) => {
     return unit.models.every((model) => {
